@@ -14,7 +14,7 @@ namespace backend.Models
         public virtual DbSet<Event> Events { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
-            => modelBuilder.ForNpgsqlUseIdentityColumns();
+            => modelBuilder.UseIdentityColumns();
 
     }
 }
